refactor: move delete template tests to core
